+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
+
+## 3.5.1 - 2018-07-13
 - Fix `implode` on `ListCollection`
 - Fix `count` on `Seq`